+++ public/results.php
@@ -10,13 +10,7 @@ use diskover\Constants;
 error_reporting(E_ALL ^ E_NOTICE);
 
 // see if there are any extra custom fields to add
-$extra_fields = [];
-for ($i=1; $i < 5; $i++) {
-    if (getCookie('field'.$i.'')) {
-        $value = (getCookie('field'.$i.'-desc')) ? getCookie('field'.$i.'-desc') : getCookie('field'.$i.'');
-        $extra_fields[getCookie('field'.$i.'')] = $value;
-    }
-}
+$extra_fields = get_extra_fields();
 
 // Grab all the custom tags from file
 $customtags = get_custom_tags();